# project/public/doc/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Cars API",
        "description": "REST API for managing cars",
        "version": "1.0.0"
    },
    "paths": {
        "/api/cars": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "List all cars",
                "operationId": "get_cars_list",
                "responses": {
                    "200": {
                        "description": "Array of cars",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Car"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Cars"
                ],
                "summary": "Create a new car",
                "operationId": "post_cars_create",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CarInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Car created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation failed",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "errors": {
                                            "type": "array",
                                            "items": {
                                                "type": "string"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/car/{id}": {
            "get": {
                "tags": [
                    "Cars"
                ],
                "summary": "Get a single car by ID",
                "operationId": "get_car_get",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Car found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Car"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/cars/{id}": {
            "delete": {
                "tags": [
                    "Cars"
                ],
                "summary": "Delete a car by ID",
                "operationId": "delete_cars_delete",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Car deleted"
                    },
                    "404": {
                        "description": "Car not found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Car not found"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Car": {
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "make": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Yaris"
                    },
                    "build_date": {
                        "type": "string",
                        "format": "date",
                        "example": "2023-03-15"
                    },
                    "colour_id": {
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            },
            "CarInput": {
                "required": [
                    "make",
                    "model",
                    "build_date",
                    "colour_id"
                ],
                "properties": {
                    "make": {
                        "type": "string",
                        "example": "Toyota"
                    },
                    "model": {
                        "type": "string",
                        "example": "Yaris"
                    },
                    "build_date": {
                        "description": "Must be within the last 4 years",
                        "type": "string",
                        "format": "date",
                        "example": "2023-03-15"
                    },
                    "colour_id": {
                        "description": "1=red, 2=blue, 3=white, 4=black",
                        "type": "integer",
                        "example": 1
                    }
                },
                "type": "object"
            }
        }
    },
    "tags": [
        {
            "name": "Cars"
        }
    ]
}